- added CardCategory, CardCategoryRepository and CardCategoryID (for @IdClass in CardCategory - added in Card and Category the CardCategory mapping - implemented in CardCategoryRepository findAllByCardId and findAllByCategoryId - implemented in DeskService getFilteredByCategory - changed in CardRepository Long to Integer

+++ src/main/java/de/thb/anylearn/repository/CardRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-@RepositoryDefinition(domainClass = Card.class, idClass = Long.class)
-public interface CardRepository extends CrudRepository<Card, Long> {
+@RepositoryDefinition(domainClass = Card.class, idClass = Integer.class)
+public interface CardRepository extends CrudRepository<Card, Integer> {
 
 
     @Query("SELECT c FROM Card c WHERE c.folder.id = :folderId")
